# Projeto_2_Biblioteca/dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"BasicResponseDto": {
				"properties": {
					"message": {
						"type": "string"
					},
					"object": {}
				},
				"required": [
					"message",
					"object"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"PessoaRequestDto": {
				"properties": {
					"nome": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"nome",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PessoaEntity": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"nome": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"nome",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LivroRequestDto": {
				"properties": {
					"titulo": {
						"type": "string"
					},
					"autor": {
						"type": "string"
					},
					"categoriaId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"titulo",
					"autor",
					"categoriaId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LivroEntity": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"titulo": {
						"type": "string"
					},
					"autor": {
						"type": "string"
					},
					"categoriaId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"titulo",
					"autor",
					"categoriaId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "API Documentação",
		"version": "1.0.0",
		"description": "Documentação da API usando Swagger"
	},
	"paths": {
...
		"/pessoa/Cadastrar/Pessoa": {
			"post": {
				"operationId": "CadastrarPessoa",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Pessoa"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PessoaRequestDto"
							}
						}
					}
				}
			}
		},
		"/pessoa/Atualizar/Pessoa": {
			"put": {
				"operationId": "AtualizarPessoa",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Pessoa"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PessoaEntity"
							}
						}
					}
				}
			}
		},
		"/pessoa/Deletar/Pessoa": {
			"delete": {
				"operationId": "DeletePessoa",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Pessoa"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "param",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/pessoa/filtrar/Pessoa": {
			"get": {
				"operationId": "FiltrarPessoa",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Pessoa"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "param",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/pessoa/Listar/Todas/as/Pessoas": {
			"get": {
				"operationId": "ListarTodasPessoas",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Pessoa"
				],
				"security": [],
				"parameters": []
			}
		},
		"/livro/Cadastrar/Livro": {
			"post": {
				"operationId": "CadastrarLivro",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Livro"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LivroRequestDto"
							}
						}
					}
				}
			}
		},
		"/livro/Atualizar/Livro": {
			"put": {
				"operationId": "AtualizarLivro",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Livro"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LivroEntity"
							}
						}
					}
				}
			}
		},
		"/livro/Deletar/Livro": {
			"delete": {
				"operationId": "DeleteLivro",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Livro"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "param",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/livro/filtrar/Livro": {
			"get": {
				"operationId": "FiltrarLivro",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Livro"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "param",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/livro/Listar/Todos/os/Livros": {
			"get": {
				"operationId": "ListarTodosLivros",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Livro"
				],
				"security": [],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3040/api",
			"description": "Servidor de Desenvolvimento"
		}
	]
}